% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rain_ruhleben.R
\name{add_average_rain_columns}
\alias{add_average_rain_columns}
\title{Add Average Rain Columns to Rain Data}
\usage{
add_average_rain_columns(rain)
}
\arguments{
\item{rain}{data frame containing rain data columns}
}
\description{
Add Average Rain Columns to Rain Data
}
