% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_and_clean}
\alias{bind_and_clean}
\title{Row-Bind and Clean Time Series in Data Frames}
\usage{
bind_and_clean(x)
}
\arguments{
\item{x}{list of data frames each of which contains a column \code{DateTime}}
}
\description{
Row-Bind and Clean Time Series in Data Frames
}
