% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_database_file_identity.R
\name{check_identity_fst_csv}
\alias{check_identity_fst_csv}
\title{Check Identity fst csv}
\usage{
check_identity_fst_csv(file_fst, file_csv, ...)
}
\arguments{
\item{file_fst}{path to fst file}

\item{file_csv}{path to csv file}

\item{...}{additional arguments passt to \code{\link{read_timeseries_csv}}}
}
\value{
message "identical or "differ"
}
\description{
Check Identity fst csv
}
