% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_structure.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Create Folder Structure}
\usage{
create_folder_structure(root, dbg = FALSE)
}
\arguments{
\item{root}{root directory}

\item{dbg}{debug (default: FALSE)}
}
\value{
create folder structure
}
\description{
Create Folder Structure
}
