% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{day_strings_until_today}
\alias{day_strings_until_today}
\title{Day Strings from Start Day Up to Today}
\usage{
day_strings_until_today(start_day)
}
\arguments{
\item{start_day}{start day as string (yyyy-mm-dd) or POSIXt object}
}
\value{
vector of character containing date strings in format "yymmdd"
}
\description{
Day Strings from Start Day Up to Today
}
