% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_path}
\alias{db_path}
\title{Path to "Database" File}
\usage{
db_path(file, root = get_root())
}
\arguments{
\item{file}{file name within application's "database" folder}

\item{root}{path to "root" directory of the application}
}
\description{
Path to "Database" File
}
