% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download a File}
\usage{
download_file(url, destfile, dbg = TRUE)
}
\arguments{
\item{url}{URL to the file to be downloaded}

\item{destfile}{full path to target file name}

\item{dbg}{show message about the downloading process if \code{TRUE}}
}
\description{
Download a File
}
