% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_day_string.R
\name{filter_for_day_string}
\alias{filter_for_day_string}
\title{Filter For Day String}
\usage{
filter_for_day_string(df, day_string)
}
\arguments{
\item{df}{data frame}

\item{day_string}{day string in ISO format "year-month-day"}
}
\value{
data frame with days smaller or equal day_string
}
\description{
Filter For Day String
}
