% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_file.R
\name{get_bwb_rain_data_columns}
\alias{get_bwb_rain_data_columns}
\title{Read Names of BWB Rain Gauges from Example Rain Data File}
\usage{
get_bwb_rain_data_columns()
}
\description{
Read Names of BWB Rain Gauges from Example Rain Data File
}
