% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_environment_variable}
\alias{get_environment_variable}
\title{Get Environment Variable Giving a Hint on Missing Variables}
\usage{
get_environment_variable(x)
}
\arguments{
\item{x}{name of environment variable}
}
\description{
Get Environment Variable Giving a Hint on Missing Variables
}
