% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_input.R
\name{get_model_input}
\alias{get_model_input}
\title{Create the Model Input List}
\usage{
get_model_input(x, logRain = TRUE, fittingData = TRUE)
}
\arguments{
\item{x}{data frame}

\item{logRain}{logical}

\item{fittingData}{logical}
}
\description{
Create the Model Input List
}
