% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_two_site_prediction.R
\name{get_prediction}
\alias{get_prediction}
\title{Get Prediction}
\usage{
get_prediction(
  model,
  d_predict,
  quantile_probs = c(0.025, 0.975, 0.5, 0.9, 0.95),
  dbg = 1,
  context = deparse(substitute(model))
)
}
\arguments{
\item{model}{model}

\item{d_predict}{d_predict}

\item{quantile_probs}{quantile_probs (default: c(0.025, 0.975, 0.5, 0.9, 0.95))}

\item{dbg}{default: 1}

\item{context}{default: deparse(substitute(model)}
}
\value{
prediction
}
\description{
Get Prediction
}
