% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quality.R
\name{get_quality}
\alias{get_quality}
\title{Get Quality}
\usage{
get_quality(prediction, context = deparse(substitute(prediction)))
}
\arguments{
\item{prediction}{prediction}

\item{context}{default: deparse(substitute(prediction))}
}
\value{
quality
}
\description{
Get Quality
}
