% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModInput.R
\name{getModInput}
\alias{getModInput}
\title{Create the Model Input List}
\usage{
getModInput(x, logRain = TRUE, fittingData = TRUE)
}
\arguments{
\item{x}{data frame}

\item{logRain}{logical}

\item{fittingData}{logical}
}
\description{
Create the Model Input List
}
