% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flows.R
\name{lines_to_flow_data}
\alias{lines_to_flow_data}
\title{Lines to Flow Data}
\usage{
lines_to_flow_data(x, columns = c("DateTime", "Flow"), switches = TRUE)
}
\arguments{
\item{x}{data frame}

\item{columns}{default: c("DateTime", "Flow")}

\item{switches}{default: TRUE}
}
\value{
Return only selected columns unless "columns" is NULL
}
\description{
Lines to Flow Data
}
