% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModInputLite.R
\name{ModInputLite}
\alias{ModInputLite}
\title{Reduce Model Input to Relevant Data}
\usage{
ModInputLite(ModInput, model)
}
\arguments{
\item{ModInput}{data frame}

\item{model}{model object}
}
\description{
Reduce Model Input to Relevant Data
}
