% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_model_input.R
\name{prepare_model_input}
\alias{prepare_model_input}
\title{Prepare Model Input Data Frame}
\usage{
prepare_model_input(rain_ruhleben, tiefwerder)
}
\arguments{
\item{rain_ruhleben}{Rain and Ruhleben data as prepared with
\code{\link{prepare_rain_ruhleben}}}

\item{tiefwerder}{Tiefwerder data as prepared with
\code{\link{prepare_tiefwerder}}}
}
\description{
Prepare Model Input Data Frame
}
