% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rain_ruhleben.R
\name{prepare_rain_ruhleben}
\alias{prepare_rain_ruhleben}
\title{Prepare Rain and Ruhleben Data}
\usage{
prepare_rain_ruhleben(combined)
}
\arguments{
\item{combined}{data frame with rain data and data from Ruhleben}
}
\description{
Prepare Rain and Ruhleben Data
}
