% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tiefwerder.R
\name{prepare_tiefwerder}
\alias{prepare_tiefwerder}
\title{Prepare Tiefwerder Data}
\usage{
prepare_tiefwerder(flows, dbg = 1)
}
\arguments{
\item{flows}{data frame with flow data from Tiefwerder and Sophienwerder}

\item{dbg}{debug level. The higher the value, the more verbose the output}
}
\description{
Prepare Tiefwerder Data
}
