% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_file.R
\name{read_bwb_file_raw}
\alias{read_bwb_file_raw}
\title{Read Rain Data from Tab-Limited File and Give Column Names}
\usage{
read_bwb_file_raw(file, column_names)
}
\arguments{
\item{file}{full path to file from which to read}

\item{column_names}{vector of names to be given to the columns in the
returned data frame}
}
\description{
Read Rain Data from Tab-Limited File and Give Column Names
}
