% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_input_csv.R
\name{read_model_input_csv}
\alias{read_model_input_csv}
\title{Read Model Input Csv}
\usage{
read_model_input_csv(file)
}
\arguments{
\item{file}{file}
}
\value{
content
}
\description{
Read Model Input Csv
}
