% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rain_from_files.R
\name{read_rain_from_files}
\alias{read_rain_from_files}
\title{Read Rain from Files}
\usage{
read_rain_from_files(files, dbg = TRUE)
}
\arguments{
\item{files}{files}

\item{dbg}{default: TRUE}
}
\value{
data frame with imported rain data
}
\description{
Read Rain from Files
}
