% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_timeseries_csv.R
\name{read_timeseries_csv}
\alias{read_timeseries_csv}
\title{Read Timeseries Csv}
\usage{
read_timeseries_csv(file, time_columns = c("tBeg", "tEnd"), switches = TRUE)
}
\arguments{
\item{file}{file}

\item{time_columns}{default: c("tBeg", "tEnd")}

\item{switches}{default: TRUE}
}
\value{
data frame with timeseries
}
\description{
Read Timeseries Csv
}
