% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_model_input.R
\name{reduce_model_input}
\alias{reduce_model_input}
\title{Reduce Model Input to Relevant Data}
\usage{
reduce_model_input(ModInput, model)
}
\arguments{
\item{ModInput}{data frame}

\item{model}{model object}
}
\description{
Reduce Model Input to Relevant Data
}
