% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_column_expected_empty}
\alias{remove_column_expected_empty}
\title{Remove Column Expected Empty}
\usage{
remove_column_expected_empty(df, column, dbg = FALSE)
}
\arguments{
\item{df}{df}

\item{column}{column}

\item{dbg}{default: FALSE}
}
\value{
data frame without "column" to be removed (if empty!)
}
\description{
Remove Column Expected Empty
}
