% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_root}
\alias{set_root}
\title{Set the Root Folder of the Flusshygiene App}
\usage{
set_root(root)
}
\arguments{
\item{root}{path to the root folder}
}
\description{
The root folder will be stored in the environment variable
"FLUSSHYGIENE_ROOT"
}
