% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rain_ruhleben.R
\name{summarise_rain_ruhleben}
\alias{summarise_rain_ruhleben}
\title{Summarise Rain Ruhleben}
\usage{
summarise_rain_ruhleben(combined)
}
\arguments{
\item{combined}{data frame "combined"}
}
\value{
add rain summary
}
\description{
Summarise Rain Ruhleben
}
