% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all_predictions.R
\name{update_all_predictions}
\alias{update_all_predictions}
\title{Update All Predictions}
\usage{
update_all_predictions(prediction_today, root = get_root())
}
\arguments{
\item{prediction_today}{prediction_today}

\item{root}{default: \link{get_root}}
}
\value{
update all predictions
}
\description{
Update All Predictions
}
