% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bwb_database.R
\name{update_bwb_database}
\alias{update_bwb_database}
\title{Download New Files and Update Local "Database"}
\usage{
update_bwb_database(root, user_pwd, dbg = TRUE)
}
\arguments{
\item{root}{path to "root" folder below which to find subfolders "downloads"
and "database"}

\item{user_pwd}{user and password string to access KWB's FTP server}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
This function checks for files that are available on KWB's FTP server and
that are not yet contained in the download folder <root>/downloads/bwb/. It
downloads new files from the FTP server into the download folder. New data
are read from the downloaded files and appended to the "database", i.e. to
two files in <root>/database/: 'rain-ruhleben.fst' (to be read with
fst::read.fst) and 'rain-ruhleben.csv'.
}
