% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_and_predict.R
\name{update_data_and_predict}
\alias{update_data_and_predict}
\title{Update Model Input Data and Predict Water Quality}
\usage{
update_data_and_predict(day_string = as.character(Sys.Date()),
  upload = FALSE)
}
\arguments{
\item{day_string}{day for which to predict the water quality, in format
yyyy-mm-dd, e.g. "2019-07-01". Default: \code{as.character(Sys.Date())}}

\item{upload}{logical. If \code{TRUE} the prediction file
\code{Vorhersage_yyyy-mm-dd.csv} is uploaded to the server of
Technologiestiftung Berlin (TSB).}
}
\description{
Update Model Input Data and Predict Water Quality
}
