% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_prediction_today.R
\name{upload_prediction_today}
\alias{upload_prediction_today}
\title{Upload Prediction Today}
\usage{
upload_prediction_today(file)
}
\arguments{
\item{file}{file}
}
\value{
upload prediction today
}
\description{
Upload Prediction Today
}
