% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_fst_file}
\alias{write_fst_file}
\title{Write Fst File}
\usage{
write_fst_file(x, file, context = deparse(substitute(x)))
}
\arguments{
\item{x}{object to be saved}

\item{file}{file}

\item{context}{default: deparse(substitute(x))}
}
\value{
write fst file
}
\description{
Write Fst File
}
