% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_environment_variable}
\alias{get_environment_variable}
\title{Get Environment Variable Giving a Hint on Missing Variables}
\usage{
get_environment_variable(x, do_stop = FALSE)
}
\arguments{
\item{x}{name of environment variable}

\item{do_stop}{logical. If \code{TRUE} an error is raised if the environment
variable is not set. Otherwise (the default), a message is shown.}
}
\description{
Get Environment Variable Giving a Hint on Missing Variables
}
