% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tiefwerder.R
\name{prepare_tiefwerder}
\alias{prepare_tiefwerder}
\title{Prepare Tiefwerder Data}
\usage{
prepare_tiefwerder(tiefwerder)
}
\arguments{
\item{tiefwerder}{data frame with data from Tiefwerder}
}
\description{
Prepare Tiefwerder Data
}
