% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bwb_file.R
\name{read_bwb_file}
\alias{read_bwb_file}
\title{Read BWB Rain Data File}
\usage{
read_bwb_file(file, column_names = get_bwb_rain_data_columns(),
  dbg = TRUE)
}
\arguments{
\item{file}{full path to rain data file}

\item{column_names}{names to be given to the columns of the returned rain
data. Defaults to \code{kwb.flusshygiene.app:::get_bwb_rain_data_columns()}}

\item{dbg}{write debug messages if \code{TRUE}}
}
\description{
Read BWB Rain Data File
}
