% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_model_input.R
\name{reduce_model_input}
\alias{reduce_model_input}
\title{Reduce Model Input to Relevant Data}
\usage{
reduce_model_input(model_input, model,
  context = deparse(substitute(model)))
}
\arguments{
\item{model_input}{data frame}

\item{model}{model object}

\item{context}{context string to appear in error messages. Default: name of
the object passed in \code{model}}
}
\description{
Reduce Model Input to Relevant Data
}
