% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_senate_database.R
\name{update_senate_database}
\alias{update_senate_database}
\title{Download New Files and Update Local "Database" of Senate's Data}
\usage{
update_senate_database(root)
}
\arguments{
\item{root}{path to "root" folder below which to find subfolders "downloads"
and "database"}
}
\description{
Download New Files and Update Local "Database" of Senate's Data
}
