% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_input_file}
\alias{write_input_file}
\title{Write Data Frame to CSV File}
\usage{
write_input_file(x, file)
}
\arguments{
\item{x}{data frame to be written to CSV file}

\item{file}{full path to target file}
}
\description{
Write Data Frame to CSV File
}
